# internal/api/domain.go
// Package api provides primitives to interact with the openapi HTTP API.
//
// Code generated by github.com/oapi-codegen/oapi-codegen/v2 version v2.5.1 DO NOT EDIT.
package api

import (
	"time"

	openapi_types "github.com/oapi-codegen/runtime/types"
)

// CreateListRequest defines model for CreateListRequest.
type CreateListRequest struct {
	Title string `json:"title"`
}

// CreateTaskRequest defines model for CreateTaskRequest.
type CreateTaskRequest struct {
	// Text Текст задачи (1-500 символов)
	Text string `json:"text"`
}

// Error defines model for Error.
type Error struct {
	// Code Машинный код ошибки
	Code string `json:"code"`

	// Details Дополнительные сведения
	Details interface{} `json:"details,omitempty"`

	// Message Человекочитаемое описание
	Message string `json:"message"`
}

// Health defines model for Health.
type Health struct {
	Status *string `json:"status,omitempty"`
}

// List defines model for List.
type List struct {
	// CreatedAt Время создания (RFC3339)
	CreatedAt *time.Time `json:"created_at,omitempty"`

	// Id Идентификатор списка
	Id openapi_types.UUID `json:"id"`

	// Title Название списка
	Title string `json:"title"`
}

// Task defines model for Task.
type Task struct {
	// Completed Статус выполнения
	Completed bool `json:"completed"`

	// CreatedAt Время создания (RFC3339)
	CreatedAt *time.Time `json:"created_at,omitempty"`

	// Id Идентификатор задачи
	Id openapi_types.UUID `json:"id"`

	// ListId Идентификатор родительского списка
	ListId openapi_types.UUID `json:"list_id"`

	// Text Текст задачи
	Text string `json:"text"`

	// UpdatedAt Время последнего обновления (RFC3339)
	UpdatedAt *time.Time `json:"updated_at,omitempty"`
}

// UpdateListRequest defines model for UpdateListRequest.
type UpdateListRequest struct {
	Title string `json:"title"`
}

// UpdateTaskRequest defines model for UpdateTaskRequest.
type UpdateTaskRequest struct {
	// Completed Статус выполнения
	Completed *bool `json:"completed,omitempty"`

	// Text Текст задачи (1-500 символов)
	Text *string `json:"text,omitempty"`
}

// Id defines model for Id.
type Id = openapi_types.UUID

// Limit defines model for Limit.
type Limit = int

// Offset defines model for Offset.
type Offset = int

// Title defines model for Title.
type Title = string

// NotFound defines model for NotFound.
type NotFound = Error

// ServerError defines model for ServerError.
type ServerError = Error

// ValidationError defines model for ValidationError.
type ValidationError = Error

// ListListsParams defines parameters for ListLists.
type ListListsParams struct {
	// Limit Максимум элементов в ответе
	Limit *Limit `form:"limit,omitempty" json:"limit,omitempty"`

	// Offset Смещение в выборке
	Offset *Offset `form:"offset,omitempty" json:"offset,omitempty"`
}

// SearchByTitleParams defines parameters for SearchByTitle.
type SearchByTitleParams struct {
	// Q Поисковый запрос по названию
	Q Title `form:"q" json:"q"`
}

// ListTasksParams defines parameters for ListTasks.
type ListTasksParams struct {
	// Limit Максимум элементов в ответе
	Limit *Limit `form:"limit,omitempty" json:"limit,omitempty"`

	// Offset Смещение в выборке
	Offset *Offset `form:"offset,omitempty" json:"offset,omitempty"`
}

// CreateListJSONRequestBody defines body for CreateList for application/json ContentType.
type CreateListJSONRequestBody = CreateListRequest

// UpdateListJSONRequestBody defines body for UpdateList for application/json ContentType.
type UpdateListJSONRequestBody = UpdateListRequest

// CreateTaskJSONRequestBody defines body for CreateTask for application/json ContentType.
type CreateTaskJSONRequestBody = CreateTaskRequest

// UpdateTaskJSONRequestBody defines body for UpdateTask for application/json ContentType.
type UpdateTaskJSONRequestBody = UpdateTaskRequest