# internal/api/server.go
// Package api provides primitives to interact with the openapi HTTP API.
//
// Code generated by github.com/oapi-codegen/oapi-codegen/v2 version v2.5.1 DO NOT EDIT.
package api

import (
	"fmt"
	"net/http"

	"github.com/labstack/echo/v4"
	"github.com/oapi-codegen/runtime"
	openapi_types "github.com/oapi-codegen/runtime/types"
)

// ServerInterface represents all server handlers.
type ServerInterface interface {
	// Получить списки
	// (GET /api/v1/lists)
	ListLists(ctx echo.Context, params ListListsParams) error
	// Создать список
	// (POST /api/v1/lists)
	CreateList(ctx echo.Context) error
	// Получить список по title
	// (GET /api/v1/lists/search)
	SearchByTitle(ctx echo.Context, params SearchByTitleParams) error
	// Удалить список
	// (DELETE /api/v1/lists/{id})
	DeleteList(ctx echo.Context, id Id) error
	// Получить список по id
	// (GET /api/v1/lists/{id})
	GetList(ctx echo.Context, id Id) error
	// Обновить title списка
	// (PATCH /api/v1/lists/{id})
	UpdateList(ctx echo.Context, id Id) error
	// Получить задачи списка
	// (GET /api/v1/lists/{listID}/tasks)
	ListTasks(ctx echo.Context, listID openapi_types.UUID, params ListTasksParams) error
	// Создать задачу в списке
	// (POST /api/v1/lists/{listID}/tasks)
	CreateTask(ctx echo.Context, listID openapi_types.UUID) error
	// Удалить задачу
	// (DELETE /api/v1/tasks/{taskID})
	DeleteTask(ctx echo.Context, taskID openapi_types.UUID) error
	// Получить задачу по ID
	// (GET /api/v1/tasks/{taskID})
	GetTask(ctx echo.Context, taskID openapi_types.UUID) error
	// Обновить задачу (частичное обновление)
	// (PATCH /api/v1/tasks/{taskID})
	UpdateTask(ctx echo.Context, taskID openapi_types.UUID) error
	// Проверка здоровья сервиса
	// (GET /health)
	Health(ctx echo.Context) error
}

// ServerInterfaceWrapper converts echo contexts to parameters.
type ServerInterfaceWrapper struct {
	Handler ServerInterface
}

// ListLists converts echo context to params.
func (w *ServerInterfaceWrapper) ListLists(ctx echo.Context) error {
	var err error

	// Parameter object where we will unmarshal all parameters from the context
	var params ListListsParams
	// ------------- Optional query parameter "limit" -------------

	err = runtime.BindQueryParameter("form", true, false, "limit", ctx.QueryParams(), &params.Limit)
	if err != nil {
		return echo.NewHTTPError(http.StatusBadRequest, fmt.Sprintf("Invalid format for parameter limit: %s", err))
	}

	// ------------- Optional query parameter "offset" -------------

	err = runtime.BindQueryParameter("form", true, false, "offset", ctx.QueryParams(), &params.Offset)
	if err != nil {
		return echo.NewHTTPError(http.StatusBadRequest, fmt.Sprintf("Invalid format for parameter offset: %s", err))
	}

	// Invoke the callback with all the unmarshaled arguments
	err = w.Handler.ListLists(ctx, params)
	return err
}

// CreateList converts echo context to params.
func (w *ServerInterfaceWrapper) CreateList(ctx echo.Context) error {
	var err error

	// Invoke the callback with all the unmarshaled arguments
	err = w.Handler.CreateList(ctx)
	return err
}

// SearchByTitle converts echo context to params.
func (w *ServerInterfaceWrapper) SearchByTitle(ctx echo.Context) error {
	var err error

	// Parameter object where we will unmarshal all parameters from the context
	var params SearchByTitleParams
	// ------------- Required query parameter "q" -------------

	err = runtime.BindQueryParameter("form", true, true, "q", ctx.QueryParams(), &params.Q)
	if err != nil {
		return echo.NewHTTPError(http.StatusBadRequest, fmt.Sprintf("Invalid format for parameter q: %s", err))
	}

	// Invoke the callback with all the unmarshaled arguments
	err = w.Handler.SearchByTitle(ctx, params)
	return err
}

// DeleteList converts echo context to params.
func (w *ServerInterfaceWrapper) DeleteList(ctx echo.Context) error {
	var err error
	// ------------- Path parameter "id" -------------
	var id Id

	err = runtime.BindStyledParameterWithOptions("simple", "id", ctx.Param("id"), &id, runtime.BindStyledParameterOptions{ParamLocation: runtime.ParamLocationPath, Explode: false, Required: true})
	if err != nil {
		return echo.NewHTTPError(http.StatusBadRequest, fmt.Sprintf("Invalid format for parameter id: %s", err))
	}

	// Invoke the callback with all the unmarshaled arguments
	err = w.Handler.DeleteList(ctx, id)
	return err
}

// GetList converts echo context to params.
func (w *ServerInterfaceWrapper) GetList(ctx echo.Context) error {
	var err error
	// ------------- Path parameter "id" -------------
	var id Id

	err = runtime.BindStyledParameterWithOptions("simple", "id", ctx.Param("id"), &id, runtime.BindStyledParameterOptions{ParamLocation: runtime.ParamLocationPath, Explode: false, Required: true})
	if err != nil {
		return echo.NewHTTPError(http.StatusBadRequest, fmt.Sprintf("Invalid format for parameter id: %s", err))
	}

	// Invoke the callback with all the unmarshaled arguments
	err = w.Handler.GetList(ctx, id)
	return err
}

// UpdateList converts echo context to params.
func (w *ServerInterfaceWrapper) UpdateList(ctx echo.Context) error {
	var err error
	// ------------- Path parameter "id" -------------
	var id Id

	err = runtime.BindStyledParameterWithOptions("simple", "id", ctx.Param("id"), &id, runtime.BindStyledParameterOptions{ParamLocation: runtime.ParamLocationPath, Explode: false, Required: true})
	if err != nil {
		return echo.NewHTTPError(http.StatusBadRequest, fmt.Sprintf("Invalid format for parameter id: %s", err))
	}

	// Invoke the callback with all the unmarshaled arguments
	err = w.Handler.UpdateList(ctx, id)
	return err
}

// ListTasks converts echo context to params.
func (w *ServerInterfaceWrapper) ListTasks(ctx echo.Context) error {
	var err error
	// ------------- Path parameter "listID" -------------
	var listID openapi_types.UUID

	err = runtime.BindStyledParameterWithOptions("simple", "listID", ctx.Param("listID"), &listID, runtime.BindStyledParameterOptions{ParamLocation: runtime.ParamLocationPath, Explode: false, Required: true})
	if err != nil {
		return echo.NewHTTPError(http.StatusBadRequest, fmt.Sprintf("Invalid format for parameter listID: %s", err))
	}

	// Parameter object where we will unmarshal all parameters from the context
	var params ListTasksParams
	// ------------- Optional query parameter "limit" -------------

	err = runtime.BindQueryParameter("form", true, false, "limit", ctx.QueryParams(), &params.Limit)
	if err != nil {
		return echo.NewHTTPError(http.StatusBadRequest, fmt.Sprintf("Invalid format for parameter limit: %s", err))
	}

	// ------------- Optional query parameter "offset" -------------

	err = runtime.BindQueryParameter("form", true, false, "offset", ctx.QueryParams(), &params.Offset)
	if err != nil {
		return echo.NewHTTPError(http.StatusBadRequest, fmt.Sprintf("Invalid format for parameter offset: %s", err))
	}

	// Invoke the callback with all the unmarshaled arguments
	err = w.Handler.ListTasks(ctx, listID, params)
	return err
}

// CreateTask converts echo context to params.
func (w *ServerInterfaceWrapper) CreateTask(ctx echo.Context) error {
	var err error
	// ------------- Path parameter "listID" -------------
	var listID openapi_types.UUID

	err = runtime.BindStyledParameterWithOptions("simple", "listID", ctx.Param("listID"), &listID, runtime.BindStyledParameterOptions{ParamLocation: runtime.ParamLocationPath, Explode: false, Required: true})
	if err != nil {
		return echo.NewHTTPError(http.StatusBadRequest, fmt.Sprintf("Invalid format for parameter listID: %s", err))
	}

	// Invoke the callback with all the unmarshaled arguments
	err = w.Handler.CreateTask(ctx, listID)
	return err
}

// DeleteTask converts echo context to params.
func (w *ServerInterfaceWrapper) DeleteTask(ctx echo.Context) error {
	var err error
	// ------------- Path parameter "taskID" -------------
	var taskID openapi_types.UUID

	err = runtime.BindStyledParameterWithOptions("simple", "taskID", ctx.Param("taskID"), &taskID, runtime.BindStyledParameterOptions{ParamLocation: runtime.ParamLocationPath, Explode: false, Required: true})
	if err != nil {
		return echo.NewHTTPError(http.StatusBadRequest, fmt.Sprintf("Invalid format for parameter taskID: %s", err))
	}

	// Invoke the callback with all the unmarshaled arguments
	err = w.Handler.DeleteTask(ctx, taskID)
	return err
}

// GetTask converts echo context to params.
func (w *ServerInterfaceWrapper) GetTask(ctx echo.Context) error {
	var err error
	// ------------- Path parameter "taskID" -------------
	var taskID openapi_types.UUID

	err = runtime.BindStyledParameterWithOptions("simple", "taskID", ctx.Param("taskID"), &taskID, runtime.BindStyledParameterOptions{ParamLocation: runtime.ParamLocationPath, Explode: false, Required: true})
	if err != nil {
		return echo.NewHTTPError(http.StatusBadRequest, fmt.Sprintf("Invalid format for parameter taskID: %s", err))
	}

	// Invoke the callback with all the unmarshaled arguments
	err = w.Handler.GetTask(ctx, taskID)
	return err
}

// UpdateTask converts echo context to params.
func (w *ServerInterfaceWrapper) UpdateTask(ctx echo.Context) error {
	var err error
	// ------------- Path parameter "taskID" -------------
	var taskID openapi_types.UUID

	err = runtime.BindStyledParameterWithOptions("simple", "taskID", ctx.Param("taskID"), &taskID, runtime.BindStyledParameterOptions{ParamLocation: runtime.ParamLocationPath, Explode: false, Required: true})
	if err != nil {
		return echo.NewHTTPError(http.StatusBadRequest, fmt.Sprintf("Invalid format for parameter taskID: %s", err))
	}

	// Invoke the callback with all the unmarshaled arguments
	err = w.Handler.UpdateTask(ctx, taskID)
	return err
}

// Health converts echo context to params.
func (w *ServerInterfaceWrapper) Health(ctx echo.Context) error {
	var err error

	// Invoke the callback with all the unmarshaled arguments
	err = w.Handler.Health(ctx)
	return err
}

// This is a simple interface which specifies echo.Route addition functions which
// are present on both echo.Echo and echo.Group, since we want to allow using
// either of them for path registration
type EchoRouter interface {
	CONNECT(path string, h echo.HandlerFunc, m ...echo.MiddlewareFunc) *echo.Route
	DELETE(path string, h echo.HandlerFunc, m ...echo.MiddlewareFunc) *echo.Route
	GET(path string, h echo.HandlerFunc, m ...echo.MiddlewareFunc) *echo.Route
	HEAD(path string, h echo.HandlerFunc, m ...echo.MiddlewareFunc) *echo.Route
	OPTIONS(path string, h echo.HandlerFunc, m ...echo.MiddlewareFunc) *echo.Route
	PATCH(path string, h echo.HandlerFunc, m ...echo.MiddlewareFunc) *echo.Route
	POST(path string, h echo.HandlerFunc, m ...echo.MiddlewareFunc) *echo.Route
	PUT(path string, h echo.HandlerFunc, m ...echo.MiddlewareFunc) *echo.Route
	TRACE(path string, h echo.HandlerFunc, m ...echo.MiddlewareFunc) *echo.Route
}

// RegisterHandlers adds each server route to the EchoRouter.
func RegisterHandlers(router EchoRouter, si ServerInterface) {
	RegisterHandlersWithBaseURL(router, si, "")
}

// Registers handlers, and prepends BaseURL to the paths, so that the paths
// can be served under a prefix.
func RegisterHandlersWithBaseURL(router EchoRouter, si ServerInterface, baseURL string) {

	wrapper := ServerInterfaceWrapper{
		Handler: si,
	}

	router.GET(baseURL+"/api/v1/lists", wrapper.ListLists)
	router.POST(baseURL+"/api/v1/lists", wrapper.CreateList)
	router.GET(baseURL+"/api/v1/lists/search", wrapper.SearchByTitle)
	router.DELETE(baseURL+"/api/v1/lists/:id", wrapper.DeleteList)
	router.GET(baseURL+"/api/v1/lists/:id", wrapper.GetList)
	router.PATCH(baseURL+"/api/v1/lists/:id", wrapper.UpdateList)
	router.GET(baseURL+"/api/v1/lists/:listID/tasks", wrapper.ListTasks)
	router.POST(baseURL+"/api/v1/lists/:listID/tasks", wrapper.CreateTask)
	router.DELETE(baseURL+"/api/v1/tasks/:taskID", wrapper.DeleteTask)
	router.GET(baseURL+"/api/v1/tasks/:taskID", wrapper.GetTask)
	router.PATCH(baseURL+"/api/v1/tasks/:taskID", wrapper.UpdateTask)
	router.GET(baseURL+"/health", wrapper.Health)

}