# openapi.yaml
openapi: 3.0.0 
definitions:
  CreateListRequest:
    properties:
      title:
        type: string
        minLength: 1
        maxLength: 200
    type: object
    required: [title]
  CreateTaskRequest:
    properties:
      text:
        type: string
        minLength: 1
        maxLength: 500
    type: object
    required: [text]
  List:
    properties:
      created_at:
        type: string
      id:
        type: string
      title:
        type: string
    type: object
  Task:
    properties:
      completed:
        type: boolean
      created_at:
        type: string
      id:
        type: string
      list_id:
        type: string
      text:
        type: string
      updated_at:
        type: string
    type: object
  UpdateListRequest:
    properties:
      title:
        type: string
        minLength: 1
        maxLength: 200
    type: object
  UpdateTaskRequest:
    properties:
      completed:
        type: boolean
      text:
        type: string
        minLength: 1
        maxLength: 500
    type: object
  ErrorResponse:
    properties:
      code:
        type: string
      details:
        type: string
      message:
        type: string
    type: object
info:
  contact: {}
  description: REST API для управления списками и задачами (CRUD, PostgreSQL)
  title: Tasks API
  version: 1.0.0
paths:
  /api/v1/lists:
    get:
      consumes:
      - application/json
      description: Возвращает список списков с пагинацией
      parameters:
      - name: limit
        in: query
        description: Лимит
        schema:
          type: integer
          minimum: 1
          maximum: 100
          default: 20
      - name: offset
        in: query
        description: Смещение
        schema:
          type: integer
          minimum: 0
          default: 0
      produces:
      - application/json
      responses:
        "200":
          description: OK
          headers:
            X-Total-Count:
              description: Общее количество списков
              type: integer
          schema:
            items:
              $ref: '#/definitions/List'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ErrorResponse'
      summary: Получить списки
      tags:
      - lists
    
    post:
      consumes:
      - application/json
      description: Создает новый список задач
      parameters:
      - description: Данные для создания списка
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/CreateListRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/List'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ErrorResponse'
      summary: Создать список
      tags:
      - lists
  /api/v1/lists/{id}:
    delete:
      consumes:
      - application/json
      description: Удаляет список по его идентификатору
      parameters:
      - description: ID списка
        in: path
        name: id
        required: true
        schema:
          type: string
          format: uuid
      produces:
      - application/json
      responses:
        "204":
          description: Удалено
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ErrorResponse'
      summary: Удалить список
      tags:
      - lists
    
    get:
      consumes:
      - application/json
      description: Возвращает список по его идентификатору
      parameters:
      - description: ID списка
        in: path
        name: id
        required: true
        schema:
          type: string
          format: uuid
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/List'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ErrorResponse'
      summary: Получить список по ID
      tags:
      - lists
    
    patch:
      consumes:
      - application/json
      description: Обновляет название списка
      parameters:
      - description: ID списка
        in: path
        name: id
        required: true
        schema:
          type: string
          format: uuid
      - description: Новое название списка
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/UpdateListRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/List'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ErrorResponse'
      summary: Обновить список
      tags:
      - lists
  /api/v1/lists/{listID}/tasks:
    get:
      consumes:
      - application/json
      description: Возвращает задачи указанного списка с пагинацией
      parameters:
      - description: ID списка
        in: path
        name: listID
        required: true
        schema:
          type: string
          format: uuid
      - name: limit
        in: query
        description: Лимит
        schema:
          type: integer
          minimum: 1
          maximum: 100
          default: 20
      - name: offset
        in: query
        description: Смещение
        schema:
          type: integer
          minimum: 0
          default: 0
      produces:
      - application/json
      responses:
        "200":
          description: OK
          headers:
            X-Total-Count:
              description: Общее количество задач
              type: integer
          schema:
            items:
              $ref: '#/definitions/Task'
            type: array
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ErrorResponse'
      summary: Получить задачи списка
      tags:
      - tasks
    
    post:
      consumes:
      - application/json
      description: Создает новую задачу в указанном списке
      parameters:
      - description: ID списка
        in: path
        name: listID
        required: true
        schema:
          type: string
          format: uuid
      - description: Данные для создания задачи
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/CreateTaskRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/Task'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ErrorResponse'
      summary: Создать задачу
      tags:
      - tasks
  /api/v1/lists/search:
    get:
      consumes:
      - application/json
      description: Возвращает списки, содержащие в названии заданную строку
      parameters:
      - description: Поисковый запрос
        in: query
        name: q
        required: true
        schema:
          type: string
          minLength: 1
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/List'
            type: array
        "400":
          description: Неверный запрос
          schema:
            $ref: '#/definitions/ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ErrorResponse'
      summary: Поиск списков по названию
      tags:
      - lists
  /api/v1/tasks/{taskID}:
    delete:
      consumes:
      - application/json
      description: Удаляет задачу по ее идентификатору
      parameters:
      - description: ID задачи
        in: path
        name: taskID
        required: true
        schema:
          type: string
          format: uuid
      produces:
      - application/json
      responses:
        "204":
          description: Удалено
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ErrorResponse'
      summary: Удалить задачу
      tags:
      - tasks
    
    get:
      consumes:
      - application/json
      description: Возвращает задачу по ее идентификатору
      parameters:
      - description: ID задачи
        in: path
        name: taskID
        required: true
        schema:
          type: string
          format: uuid
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/Task'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ErrorResponse'
      summary: Получить задачу по ID
      tags:
      - tasks
    
    patch:
      consumes:
      - application/json
      description: Обновляет описание и/или статус выполнения задачи
      parameters:
      - description: ID задачи
        in: path
        name: taskID
        required: true
        schema:
          type: string
          format: uuid
      - description: Данные для обновления задачи
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/UpdateTaskRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/Task'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ErrorResponse'
      summary: Обновить задачу
      tags:
      - tasks
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey